빠른 입장/나가기 추가

+++ vs_room_md.hpp
@@ -8,9 +8,9 @@ typedef std::shared_ptr<vs_room> vs_room_ptr;
 
 class vs_room_md {
 
-  std::map<long, vs_room_ptr> vs_rooms;
+  std::map<size_t, vs_room_ptr> vs_rooms;
   std::mutex mu;
-  size_t gen_cnt;
+  int gen_cnt;
 
   boost::asio::io_service io_service_;
   std::shared_ptr<boost::asio::io_service::work> work_ptr_;
@@ -31,12 +31,13 @@ public:
   void join_room(user_ptr user);
   void leave_room(user_ptr user);
 
-
+  /*
   void join_tmp_room(user_ptr user) {
     std::lock_guard<std::mutex> lock(mu);
     users.push_back(user);
   }
   std::vector<user_ptr> users;
+  */
 
   boost::asio::io_service& get_io_service();
 };
